Initialize ShoppingCart.mjs implementation; requires further fixes
